feat(OPS-11): Adds order placement functionality.

+++ src/main/java/org/viators/orderprocessingsystem/product/ProductT.java
@@ -30,5 +30,5 @@ public class ProductT extends BaseEntity {
     private CategoryEnum category;
 
     @Column(name = "stock_quantity")
-    private Long stockQuantity;
+    private BigDecimal stockQuantity;
 }
